added a separate widget for the button
