+++ lust/lib/views/checkinPage.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
+import 'package:lust/widgets/buttonCheck.dart';
 import 'package:lust/widgets/widgetCreator.dart';
 import 'package:lust/widgets/getTextBox.dart';
 
@@ -10,16 +11,9 @@ class CheckInOut extends StatefulWidget {
 
 class _CheckInOutState extends State<CheckInOut> {
   final String _title = "Check In/Out";
-  String _example; //PROVISIONAL
-
-  String _textButton = "Check In!";
-  MaterialColor _colorButton = Colors.green; //change once pressed the button
-  MaterialColor _splashButton = Colors.red;
-  bool _buttonState = false;
 
   @override
   Widget build(BuildContext context) {
-    print("INICIO + $_textButton");
     return Scaffold(
         appBar: AppBar(title: Text(_title)),
         body: Center(
@@ -34,60 +28,8 @@ class _CheckInOutState extends State<CheckInOut> {
                         context, "Time goal", "Be realistic!", Icons.timer),
                     TextBox.getTextBox(context, "Locker number",
                         "I know that is difficult to remember", Icons.lock),
-                    Padding(
-                        padding: EdgeInsets.only(top: 60),
-                        child:
-                            /*MaterialButton(
-                          height: 50,
-                          onPressed: onButtonPressed,
-                          child: Text(
-                            _textButton,
-                            //style: TextStyle(fontSize: 28),
-                          ),
-                          padding: EdgeInsets.all(20),
-                          color: _colorButton,
-                          splashColor: _splashButton,
-                          elevation: 5,
-                          shape: CircleBorder(
-                              side: BorderSide(
-                                  style: BorderStyle.solid,
-                                  color: Colors.black26,
-                                  width: 10)),
-                        )*/
-
-                            InkWell(
-                                onTap: onButtonPressed,
-                                child: Container(
-                                    height: 150,
-                                    width: 150,
-                                    decoration: BoxDecoration(
-                                        borderRadius:
-                                            BorderRadius.circular(150),
-                                        color: _colorButton),
-                                    child: Center(
-                                        child: Text(
-                                      _textButton,
-                                      style: TextStyle(fontSize: 28),
-                                    )))))
-                  ],
-                ))));
+                    ButtonCheck(),
+                    ]))));
   }
 
-  void onButtonPressed() {
-    setState(() {
-      print("HOP");
-
-      if (_buttonState == true) {
-        _buttonState = false;
-        _colorButton = Colors.green;
-        _splashButton = Colors.red;
-        _textButton = "Check In!";
-      } else {
-        _buttonState = true;
-        _colorButton = Colors.red;
-        _splashButton = Colors.green;
-        _textButton = "Check out";
-      }
-    });
-  }
 }
